Add files via upload
The project's base file, from which all other project-related modules are called, serves as its starting point. Since we used Alexnet for our project to train images, this file calls the Alexnet module. This file also includes the graph that must be plotted; the emotion with the highest % will be selected as the output depending on the graph.